[SQUASH] Settings: Gestures: Add option to cycle through ringer modes [2/2]
[jhonboy121]: adapt to A13

Signed-off-by: rituj <ritujbeniwal@gmail.com>
Change-Id: If4a387a9f8a465d33b25b87b0360fa20f1cee7de
Signed-off-by: jhonboy121 <alfredmathew05@gmail.com>

This is the commit message #2:

Settings: Choose which ringer modes to cycle with the gesture [2/2]

User can still select a single mode, which disables cycling and just
applies the desired mode.
Nuke the unit test for now as I'm lazy. Will consider writing a new one in
the future.

[jhonboy121]: adapt to A13

Signed-off-by: jhonboy121 <alfredmathew05@gmail.com>

This is the commit message #3:

Settings: improve PreventRingingGesturePreferenceController

* made many variables private that aren't used anywhere else
* use String.join() to flatten the list
* simplified SettingsObserver

Signed-off-by: jhonboy121 <alfredmathew05@gmail.com>

Co-authored-by: Ido Ben-Hur <idoybh2@gmail.com>
Co-authored-by: jhonboy121 <alfredmathew05@gmail.com>
Signed-off-by: jhonboy121 <alfredmathew05@gmail.com>